changes to product component
