+++ bonsite/src/app/page.tsx
@@ -1,7 +1,13 @@
+"use client"
+
+import Product from '@/components/ProductList';
+
+
 export default function Page() {
     return <>
 
-    
+      <Product></Product>
+
 
     </>;
